Combining files.
