--- src/async-api/interface/index.ts
@@ -1,3 +1,0 @@
-export * from "./commons";
-export * from "./server";
-export * from "./socket-io-binding";
