--- src/async-api/interface/server.ts
@@ -1,11 +1,0 @@
-import type { ServerObject } from "openapi3-ts/oas31";
-import { SocketIOServerBinding } from "./socket-io-binding";
-import { AsyncServerVariableObject, Protocol, SecurityObject } from "./commons";
-
-export interface AsyncServerObject extends Omit<ServerObject, "variables"> {
-  variables?: Record<string, AsyncServerVariableObject>;
-  protocol: Protocol;
-  protocolVersion?: string;
-  security?: SecurityObject[];
-  bindings?: Partial<Record<Protocol, SocketIOServerBinding>>;
-}
